--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -108,6 +108,9 @@ zstd = '=0.11.1'
 fwdansi = "=1.1.0"
 winapi = { version = "=0.3.9", features = ["knownfolders", "mswsock", "objbase", "shlobj", "tlhelp32", "winbase", "winerror", "winsock2"] }
 
+[target.'cfg(not(target_env = "msvc"))'.dependencies]
+tikv-jemallocator = "0.5"
+
 [dev-dependencies]
 csv = "1.1.6"
 deno_bench_util = { version = "0.54.0", path = "../bench_util" }