@lucacasonato 2022-07-08
Best out of three runs: 70.1krps (main) vs 71.0krps (this pr)

```
# main
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   151.37us  120.56us   3.48ms   98.09%
    Req/Sec    35.26k     1.95k   39.76k    77.72%
  Latency Distribution
     50%  133.00us
     75%  155.00us
     90%  166.00us
     99%  610.00us
  708532 requests in 10.10s, 74.33MB read
Requests/sec:  70154.09
Transfer/sec:      7.36MB
```

```
# this pr
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   151.82us  145.90us   5.25ms   98.22%
    Req/Sec    35.71k     1.89k   39.96k    84.16%
  Latency Distribution
     50%  130.00us
     75%  153.00us
     90%  164.00us
     99%  657.00us
  717784 requests in 10.10s, 75.30MB read
Requests/sec:  71070.54
Transfer/sec:      7.46MB
```